skill_recommendation/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
textract
python-multipart